html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-grid-flex-multicol.html

<!doctype html>
<title>legend and flexbox, grid & multicol</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<link rel=stylesheet href=resources/fieldset-default-style.css>
<style>
.fieldset-default-style { -webkit-appearance: fieldset; }
.legend-default-style { -webkit-appearance: fieldset-legend; }
legend, .legend-default-style { width: 200px; background: silver }
#flex, #appearance-flex { display: flex; }
#inline-flex, #appearance-inline-flex { display: inline-flex; }
#grid, #appearance-grid { display: grid; }
#inline-grid, #appearance-inline-grid { display: inline-grid; }
#grid, #inline-grid, #appearance-grid, #appearance-inline-grid { grid-template-columns: auto auto }
#multicol, #appearance-multicol { columns: 2; }
</style>
<fieldset><legend id=ref>12</legend></fieldset>
<fieldset><legend id=flex><div>1</div><div>2</div></legend></fieldset>
<fieldset><legend id=inline-flex><div>1</div><div>2</div></legend></fieldset>
<fieldset><legend id=grid><div>1</div><div>2</div></legend></fieldset>
<fieldset><legend id=inline-grid><div>1</div><div>2</div></legend></fieldset>
<fieldset><legend id=multicol><div>1</div><div>2</div></legend></fieldset>
<div class=fieldset-default-style>
  <div class=legend-default-style id=appearance-flex><div>1</div><div>2</div></legend>
</div>
<div class=fieldset-default-style>
  <div class=legend-default-style id=appearance-inline-flex><div>1</div><div>2</div></div>
</div>
<div class=fieldset-default-style>
  <div class=legend-default-style id=appearance-grid><div>1</div><div>2</div></div>
</div>
<div class=fieldset-default-style>
  <div class=legend-default-style id=appearance-inline-grid><div>1</div><div>2</div></div>
</div>
<div class=fieldset-default-style>
  <div class=legend-default-style id=appearance-multicol><div>1</div><div>2</div></div>
</div>
<script>
 const ref = document.getElementById('ref');
 for (const id of ["flex", "inline-flex", "grid", "inline-grid", "multicol",
                   "appearance-flex", "appearance-inline-flex", "appearance-grid",
                   "appearance-inline-grid", "appearance-multicol"]) {
   test(() => {
     const elm = document.getElementById(id);
     assert_equals(elm.offsetHeight, ref.offsetHeight, 'offsetHeight');
     const expectedDisplay = id.replace('appearance-', '');
     if (expectedDisplay !== "multicol") {
       assert_equals(getComputedStyle(elm).display, expectedDisplay, 'display');
     }
   }, id);
 }
</script>
